+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ## Other
 
+- Include contents of custom assets `metadata.yaml` in `--diagnostics`. See #2107 (@Enselic)
+
 ## Syntaxes
 
 - Mapped clang-format config file (.clang-format) to YAML syntax (@TruncatedDinosour)